chore: transfer hash to group tests
G1 tests remain the same
G2 tests' values need to be adjustted, because we skip if P::BaseField
== zero (which is equivalent in the Fp2 case for both c0 and c1 to be
zero), while previously we'd skip if either F2.c0 or F2.c1 was zero
